+++ app/_components/ui/PhotoFrame.tsx
@@ -1,3 +1,4 @@
+import Image from "next/image";
 export function PhotoFrame({
   src,
   alt = "",
@@ -52,11 +53,12 @@ export function PhotoFrame({
           boxShadow: "var(--shadow-photo)",
         }}
       >
-        {/* eslint-disable-next-line @next/next/no-img-element */}
-        <img
+        <Image
           src={src}
           alt={alt}
-          style={{ width: "100%", height: "100%", objectFit: "cover", display: "block" }}
+          fill
+          sizes="(max-width: 800px) 90vw, 380px"
+          style={{ objectFit: "cover" }}
         />
       </div>
     </div>